Modifications to add Smart product interface
The ProductionRecipeDTO is used to receive the information of the production recipe.
The WorkflowCreator is where the processing of the recipe is done, run in parallel when a smart product arrives and transmit its recipe.
The WorkflowResource is expanded with more services to interact with the smart product, and the WorkflowManager to request orchestration.